kopf/requirements.txt
kopf==1.38.0
pykube-ng==23.6.0